Create HearingHelpers.ts
